move triad chords to triad package
