+++ src/main/scala/chord/triad/DiatonicTriad.scala
@@ -1,5 +1,6 @@
-package chord
+package chord.triad
 
+import chord.Chord
 import key.{Key, MajorKey, MinorKey}
 import note.Note
